(PC-19953) refactor(maintenance): make features/maintenance following the code structure
