+++ src/features/errors/pages/ScreenErrorProvider.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 
 import { ForceUpdate } from 'features/forceUpdate/ForceUpdate'
 import { useMustUpdateApp } from 'features/forceUpdate/useMustUpdateApp'
-import { MaintenanceErrorPage } from 'features/maintenance/MaintenanceErrorPage'
-import { useMaintenance } from 'features/maintenance/useMaintenance'
+import { useMaintenance } from 'features/maintenance/helpers/useMaintenance'
+import { MaintenanceErrorPage } from 'features/maintenance/pages/MaintenanceErrorPage'
 import { MAINTENANCE } from 'libs/firebase/firestore/maintenance'
 import { ScreenError } from 'libs/monitoring/errors'
